Quim

+++ models/m_area.js
@@ -5,8 +5,8 @@
 module.exports = function(sequelize, DataTypes) {
     var Area = sequelize.define('Area', {
         nom : DataTypes.STRING(1024),
-        preuMinut: DataTypes.INTEGER,
-        maxMinuts: DataTypes.INTEGER
+        preuMinut: DataTypes.FLOAT,
+        maxMinuts: DataTypes.FLOAT
     }, {
         classMethods : {
